requirements.txt
asgiref==3.7.2
Django==4.2.17
django-cors-headers==4.4.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
Pillow==10.1.0
pytz==2025.2
tzdata==2025.2
sqlparse==0.4.4
PyJWT==2.10.1